Added list of active neurons and synapses to process at each timestep, added brain logic to process those at each timestep, and reformatted firing logic to accommodate the new active lists.
git-svn-id: https://elysia.svn.sourceforge.net/svnroot/elysia/trunk@830 e0ac87d7-b42b-0410-9034-8248a05cb448

+++ brain/CellComponent.cpp
@@ -26,9 +26,9 @@ void CellComponent::syncBranchDensity(float parentRandomBranchDeterminer, float 
     }else {
        branchiness=floor(branchiness);
     }
-    size_t maxDepth=floor(parentRandomDepthDeterminer);
+    size_t maxDepth=(size_t)floor(parentRandomDepthDeterminer);
     if (parentRandomDepthDeterminer>treeDepth-floor(treeDepth))
-      maxDepth=ceil(treeDepth);
+      maxDepth=(size_t)ceil(treeDepth);
     if (depth>=(int)maxDepth) {
         branchiness=0;
     }
